- review changes:  - make enum easier  - make Int64 converting where it is needed

+++ src/xcode/ENA/ENA/Source/Extensions/Date+UTCMidnightUnixTimestamp.swift
@@ -6,7 +6,7 @@ import Foundation
 
 extension Date {
 
-	var unixTimestamp: Int64 {
-		return Int64(self.timeIntervalSince1970)
+	var unixTimestamp: Int {
+		return Int(self.timeIntervalSince1970)
 	}
 }
