+++ src/xcode/ENA/ENA/Source/Services/PPAnalyticsSubmitter/PPAnalyticsSubmitter.swift
@@ -309,6 +309,6 @@ final class PPAnalyticsSubmitter: PPAnalyticsSubmitting {
 			Log.warning("mostRecentDate is nil", log: .ppa)
 			return -1
 		}
-		return date.unixTimestamp
+		return Int64(date.unixTimestamp)
 	}
 }
